+++ test/compilationTests/MultiSigWallet/TestToken.sol
@@ -27,7 +27,7 @@ contract TestToken {
         returns (bool success)
     {
         if (balances[msg.sender] < _value) {
-            throw;
+            revert();
         }
         balances[msg.sender] -= _value;
         balances[_to] += _value;
@@ -40,7 +40,7 @@ contract TestToken {
         returns (bool success)
     {
         if (balances[_from] < _value || allowed[_from][msg.sender] < _value) {
-            throw;
+            revert();
         }
         balances[_to] += _value;
         balances[_from] -= _value;