+++ test/compilationTests/zeppelin/MultisigWallet.sol
@@ -61,7 +61,7 @@ contract MultisigWallet is Multisig, Shareable, DayLimit {
       emit SingleTransact(msg.sender, _value, _to, _data);
       // yes - just execute the call.
       if (!_to.call.value(_value)(_data)) {
-        throw;
+        revert();
       }
       return 0;
     }
@@ -83,7 +83,7 @@ contract MultisigWallet is Multisig, Shareable, DayLimit {
   function confirm(bytes32 _h) onlymanyowners(_h) public returns (bool) {
     if (txs[_h].to != address(0)) {
       if (!txs[_h].to.call.value(txs[_h].value)(txs[_h].data)) {
-        throw;
+        revert();
       }
       emit MultiTransact(msg.sender, _h, txs[_h].value, txs[_h].to, txs[_h].data);
       delete txs[_h];