+++ test/compilationTests/zeppelin/ownership/Ownable.sol
@@ -24,7 +24,7 @@ contract Ownable {
    */
   modifier onlyOwner() {
     if (msg.sender != owner) {
-      throw;
+      revert();
     }
     _;
   }